% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtClean.R
\name{jtClean}
\alias{jtClean}
\title{Feature labelling}
\usage{
jtClean(table, date, sales, product, min_date, max_date)
}
\arguments{
\item{table}{whole data}

\item{date}{string with the original name of the column containing product names}
}
\value{
Returns table with proper data structure
}
\description{
Function to set data structure
}
\keyword{collection}
\keyword{data}
