% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtLotto.R
\name{jtEnhanceLotto}
\alias{jtEnhanceLotto}
\title{Function to enhance Lotto data. Using \code{tk_augment_timeseries_signature()}}
\usage{
jtEnhanceLotto(table)
}
\arguments{
\item{table}{raw Lotto data}
}
\value{
Returns table with additional features describing time.
}
\description{
Function to enhance Lotto data. Using \code{tk_augment_timeseries_signature()}
}
\examples{
jtGetLotto()$data \%>\% jtEnhanceLotto()

}
