% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtEnhancets.R
\name{jtEnhancets}
\alias{jtEnhancets}
\title{Enhancing timeseries}
\usage{
jtEnhancets(table, sales = "Sprzedaz")
}
\arguments{
\item{table}{whole data}

\item{sales}{name of the column containing sales values}
}
\value{
Returns table with additional columns
}
\description{
Function to create multiple features using past values.
}
\keyword{data}
\keyword{processing}
