% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtLotto.R
\name{jtGetDesiredLotto}
\alias{jtGetDesiredLotto}
\title{Function to extract desired values using quantile()}
\usage{
jtGetDesiredLotto(table)
}
\arguments{
\item{table}{preprocessed Lotto data.
sum \verb{sum of the numbers sum(from 1 to 6)}
sum_of_squares \verb{sum of squared numbers sum(from 1 to 2401)}
distance_index \code{sum_of_squares / sum}}
}
\value{
Returns filtered table containing only
the combinations selected by the requirements.
}
\description{
Function to extract desired values using quantile()
}
