% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtLotto.R
\name{jtGetLotto}
\alias{jtGetLotto}
\title{Function to read the data from the file / only shiny}
\usage{
jtGetLotto(path = "http://www.mbnet.com.pl/dl.txt")
}
\arguments{
\item{path}{web page storing results}
}
\value{
Returns table with winning numbers with dates
}
\description{
Function to read the data from the file / only shiny
}
