% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtRaw.R
\name{jtRaw}
\alias{jtRaw}
\title{Sales data collection}
\usage{
jtRaw(file = input$raw_file)
}
\arguments{
\item{file}{list with stores datapath element containing the path to the file}
}
\value{
Returns table with original data structure
}
\description{
Function to read the data from the file / only shiny
}
\keyword{collection}
\keyword{data}
