% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtSeasons.R
\name{jtSeasons}
\alias{jtSeasons}
\title{Season Summary}
\usage{
jtSeasons(table)
}
\arguments{
\item{table}{preprocessed data}
}
\value{
Returns summarised table
}
\description{
Function to summarise preprocessed timeseries.
}
\examples{

clean <- table \%>\%
 jtClean(
  date = input$data,
  sales = input$sales,
  product = FALSE,
  min_date,
  max_date
  )

totals <- clean \%>\%
 jtTotals()

seasons <- totals \%>\%
 jtSeasons()

}
\keyword{data}
\keyword{processing}
