% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtTheme.R
\name{jtTheme}
\alias{jtTheme}
\title{Theme Set}
\usage{
jtTheme(
  base_size = 13,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20
)
}
\arguments{
\item{table}{preprocessed data}
}
\value{
Returns customized graph
}
\description{
Function to custiomize ggplots. Dark layout
}
\keyword{data}
\keyword{visualization}
