% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtTotals.R
\name{jtTotals}
\alias{jtTotals}
\title{Totals calculation}
\usage{
jtTotals(clean)
}
\arguments{
\item{clean}{preprocessed data}
}
\value{
Returns table with additional fixed sales value
}
\description{
Function to avoid aggregation if multiple columns are provided
}
\keyword{data}
\keyword{processing}
