% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtXGBoost.R
\name{jtXGBoost.lotto}
\alias{jtXGBoost.lotto}
\title{Function to run XGBoost for preprocessed Lotto data. Allows to
use as a target \code{dependent_var} for formula construction. The rest of the
variables are explanatory.
Using \code{xgboost}, \code{read.csv("lotto.csv")}}
\usage{
jtXGBoost.lotto(model, dependent_var)
}
\arguments{
\item{model}{preprocessed data with selected features}

\item{dependent_var}{character with the name of the variable}
}
\value{
Returns table with predictions.
}
\description{
Function to run XGBoost for preprocessed Lotto data. Allows to
use as a target \code{dependent_var} for formula construction. The rest of the
variables are explanatory.
Using \code{xgboost}, \code{read.csv("lotto.csv")}
}
\examples{
 jtXGBoost.lotto(model1, "number1")$forecasts

}
