% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtXGBoost.R
\name{jtXGBoost}
\alias{jtXGBoost}
\title{Function to run XGBoost for preprocessed sales data.}
\usage{
jtXGBoost(clean)
}
\arguments{
\item{table}{preprocessed data}
}
\value{
Returns table with predictions.
}
\description{
Flow prepared for enhanced data structure.
Using \code{xgboost} and \code{mice}
}
\examples{

clean <- table \%>\%
jtClean(
 date = input$data,
 sales = input$sales,
 product = FALSE,
 min_date,
 max_date
 )

xgbData <- clean \%>\%
jtXGBoost()

}
