% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtClean.R
\name{jtClean}
\alias{jtClean}
\title{Table structure recognition}
\usage{
jtClean(table, date, sales, product, min_date, max_date)
}
\arguments{
\item{table}{whole data}

\item{date}{string with the original name of the column containing product names}
}
\value{
Returns table with proper data structure
}
\description{
Function to set data structure
}
\examples{

jtClean(
 jtRaw(input$raw_file),
 date = input$data,
 sales = input$sales,
 product = FALSE,
 min_date = input$date[1],
 max_date = input$date[2]
 )

}
\keyword{collection}
\keyword{data}
