% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtLotto.R
\name{jtGetLotto}
\alias{jtGetLotto}
\title{Function to read the data from \verb{http://www.mbnet.com.pl/dl.txt}}
\usage{
jtGetLotto(path = "http://www.mbnet.com.pl/dl.txt")
}
\arguments{
\item{path}{web page storing results}
}
\value{
A list with the following data frames
\itemize{
\item $data - original Lotto numbers
\item $helper$prob_tables - historical probablities
for numbers selected (from 1 to 49)
}
}
\description{
Function to read the data from \verb{http://www.mbnet.com.pl/dl.txt}
}
\examples{
 jtGetLotto(path = "http://www.mbnet.com.pl/dl.txt")

}
