% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtSeasons.R
\name{jtSeasons}
\alias{jtSeasons}
\title{Season Summary}
\usage{
jtSeasons(table)
}
\arguments{
\item{table}{preprocessed data}
}
\value{
Returns summarised table
}
\description{
Function to summarise preprocessed timeseries.
}
\keyword{data}
\keyword{processing}
