% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtLotto.R
\name{predictLotto}
\alias{predictLotto}
\title{Function to run multiple Lotto models. Using \code{jtXGBoost.lotto()}}
\usage{
predictLotto(data)
}
\arguments{
\item{table}{preprocessed Lotto data}
}
\value{
Returns table with Lotto predictions.
}
\description{
Function to run multiple Lotto models. Using \code{jtXGBoost.lotto()}
}
